chore(useChildSlots): move useChildSlots from shared to hooks
